apperently it can't use numeric either, back to float for Cafes migration

+++ migrations/20180908091440_Cafes.js
@@ -2,8 +2,8 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('Cafes', table => {
     table.increments('id').primary;
     table.string('cafe_name');
-    table.numeric('latitude');
-    table.numeric('longitude');
+    table.float('latitude');
+    table.float('longitude');
     table.string('website');
   })
 }
